feat: autoriser les quantités décimales (0.5 sac, 1.5 kg, etc.)

+++ src/main/java/com/example/dijasaliou/controller/StockController.java
@@ -124,7 +124,7 @@ public class StockController {
     @GetMapping("/verifier")
     public ResponseEntity<Map<String, Object>> verifierStock(
             @RequestParam String nomProduit,
-            @RequestParam Integer quantite) {
+            @RequestParam Double quantite) {
 
         boolean disponible = stockService.verifierStockDisponible(nomProduit, quantite);
         StockDto stock = null;
